FrameLimit fix

+++ src/FrameLimit.cpp
@@ -10,8 +10,8 @@ class FrameLimitSA : public SimpleAdjuster
         const char* GetLimitName() { return GetGVM().IsSA()? "FrameLimit" : nullptr; }
         void ChangeLimit(int, const std::string& value) 
         { 
-            injector::WriteMemory(0x619626, std::stof(value), true); 
-            injector::WriteMemory(0xC1704C, std::stof(value), true);
+            injector::WriteMemory(0x619626, std::stoi(value), true); 
+            injector::WriteMemory(0xC1704C, std::stoi(value), true);
         }
         
 } FrameLimitSA;
@@ -22,8 +22,8 @@ public:
     const char* GetLimitName() { return GetGVM().IsVC() ? "FrameLimit" : nullptr; }
     void ChangeLimit(int, const std::string& value) 
     { 
-        injector::WriteMemory(0x602D68, std::stof(value), true); 
-        injector::WriteMemory(0x9B48EC, std::stof(value), true);
+        injector::WriteMemory(0x602D68, std::stoi(value), true); 
+        injector::WriteMemory(0x9B48EC, std::stoi(value), true);
     }
 
 } FrameLimitVC;
@@ -34,8 +34,8 @@ public:
     const char* GetLimitName() { return GetGVM().IsIII() ? "FrameLimit" : nullptr; }
     void ChangeLimit(int, const std::string& value) 
     { 
-        injector::WriteMemory(0x584C78, std::stof(value), true); 
-        injector::WriteMemory(0x8F4374, std::stof(value), true);
+        injector::WriteMemory(0x584C78, std::stoi(value), true); 
+        injector::WriteMemory(0x8F4374, std::stoi(value), true);
     }
 
 } FrameLimitIII;
